requirements.txt
# LLM & RAG
langchain>=0.1.0,<0.2.0
langchain-community>=0.0.10

# Document Processing
pypdf>=3.17.0
python-docx>=0.8.11
openpyxl>=3.1.0,<3.2.0

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0

# Vector DB (Local - NO API KEY NEEDED!)
chromadb>=0.4.0

# Embeddings (HuggingFace - FREE!)
sentence-transformers>=2.2.0

# Frontend
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
python-dateutil>=2.8.0
